Move bucket components into separate screens folder
